+++ src/components/burger-constructor/components/burger-empty/burger-empty.module.css
@@ -3,6 +3,7 @@
   flex-direction: column;
   align-items: center;
   justify-content: center;
+  width: 600px;
 }
 
 .burger {
